+++ mycodo/inputs/winsen_zh03b.py
@@ -80,6 +80,16 @@ INPUT_INFORMATION = {
             'constraints_pass': constraints_pass_positive_value,
             'name': lazy_gettext('Fan On Duration'),
             'phrase': lazy_gettext('How long to turn the fan on (seconds) before acquiring measurements')
+        },
+        {
+            'id': 'number_measurements',
+            'type': 'integer',
+            'default_value': 3,
+            'constraints_pass': constraints_pass_positive_value,
+            'name': lazy_gettext('Number of Measurements'),
+            'phrase': lazy_gettext(
+                'How many measurements to acquire. If more than 1 are acquired that '
+                'are less than 1001, the average of the measurements will be stored.')
         }
     ]
 }
@@ -96,6 +106,7 @@ class InputModule(AbstractInput):
         # Initialize custom options
         self.fan_modulate = None
         self.fan_seconds = None
+        self.number_measurements = None
         # Set custom options
         self.setup_custom_options(
             INPUT_INFORMATION['custom_options'], input_dev)
@@ -121,6 +132,7 @@ class InputModule(AbstractInput):
                         timeout=10
                     )
                     self.ser.flushInput()
+                    self.set_qa()
 
                     if not self.fan_modulate:
                         self.dormant_mode('run')
@@ -140,12 +152,14 @@ class InputModule(AbstractInput):
         if not self.serial_device:  # Don't measure if device isn't validated
             return None
 
+        pm_1_0 = []
+        pm_2_5 = []
+        pm_10_0 = []
+
         self.return_dict = measurements_dict.copy()
 
-        self.logger.debug("Reading sample")
-
         if self.fan_modulate and not self.fan_is_on:
-            # Allow the fan to run before querying sensor
+            # Allow the fan to run for a duration before querying sensor
             self.dormant_mode('run')
             start_time = time.time()
             while (self.running and
@@ -153,26 +167,50 @@ class InputModule(AbstractInput):
                 time.sleep(0.01)
 
         # Acquire measurements
-        pm_1_0, pm_2_5, pm_10_0 = self.qa_read_sample()
+        for i in range(self.number_measurements):
+            self.logger.debug("Acquiring measurement {}".format(i + 1))
+            pm_1_0_tmp, pm_2_5_tmp, pm_10_0_tmp = self.qa_read_sample()
+            self.logger.debug(
+                "Measurements: PM1 {}, PM2.5 {}, PM10 {}".format(
+                    pm_1_0_tmp, pm_2_5_tmp, pm_10_0_tmp))
 
-        if pm_1_0 > 1000:
-            pm_1_0 = 1001
-            self.logger.error("PM1 measurement out of range (over 1000 ug/m^3)")
-        if pm_2_5 > 1000:
-            pm_2_5 = 1001
-            self.logger.error("PM2.5 measurement out of range (over 1000 ug/m^3)")
-        if pm_10_0 > 1000:
-            pm_10_0 = 1001
-            self.logger.error("PM10 measurement out of range (over 1000 ug/m^3)")
+            if pm_1_0_tmp > 1000:
+                self.logger.debug(
+                    "PM1 measurement out of range (over 1000 ug/m^3): {}. "
+                    "Discarding.".format(pm_1_0_tmp))
+            else:
+                pm_1_0.append(pm_1_0_tmp)
 
-        if self.is_enabled(0):
-            self.value_set(0, pm_1_0)
+            if pm_2_5_tmp > 1000:
+                self.logger.debug(
+                    "PM2.5 measurement out of range (over 1000 ug/m^3): {}. "
+                    "Discarding.".format(pm_2_5_tmp))
+            else:
+                pm_2_5.append(pm_2_5_tmp)
 
-        if self.is_enabled(1):
-            self.value_set(1, pm_2_5)
+            if pm_10_0_tmp > 1000:
+                self.logger.debug(
+                    "PM10 measurement out of range (over 1000 ug/m^3): {}. "
+                    "Discarding.".format(pm_10_0_tmp))
+            else:
+                pm_10_0.append(pm_10_0_tmp)
 
-        if self.is_enabled(2):
-            self.value_set(2, pm_10_0)
+            time.sleep(0.1)
+
+        # Store measurements
+        if len(pm_1_0) < 1 or len(pm_2_5) < 1 or len(pm_10_0) < 1:
+            self.logger.debug(
+                "Error: Each particle size must have at least 1 valid "
+                "measurement to store.")
+        else:
+            if self.is_enabled(0):
+                self.value_set(0, sum(pm_1_0) / len(pm_1_0))
+
+            if self.is_enabled(1):
+                self.value_set(1, sum(pm_2_5) / len(pm_2_5))
+
+            if self.is_enabled(2):
+                self.value_set(2, sum(pm_10_0) / len(pm_10_0))
 
         # Turn the fan off
         if self.fan_modulate: