+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fix graph Widget custom color issues ([#760](https://github.com/kizniche/mycodo/issues/760))
  - Fix PWM Trigger Functions reacting to 0 % duty cycle being set ([#761](https://github.com/kizniche/mycodo/issues/761))
  - Fix KeyError if missing options when saving Input
+ - Fix ZH03B Input: add repeat measurement option and discard erroneous measurements
 
 ### Features
 
@@ -15,7 +16,7 @@
 
 ### Miscellaneous
 
- - Improve Conditional Controller code generation
+ - Upon controller activation, generate Input and Conditional code files if they don't exist
  - Update Werkzeug to 1.0.1 ([#742](https://github.com/kizniche/mycodo/issues/742)), Flask-RESTX to 0.2.0, alembic to 1.4.2, pyro5 to 5.8, SQLAlchemy to 1.3.15, distro to 1.5.0,